refactor(test): 테스트를 위한 user 클래스 분리
