+++ jdbc/src/test/java/nextstep/jdbc/User.java
@@ -1,0 +1,24 @@
+package nextstep.jdbc;
+
+public class User {
+
+    private long id;
+    private String username;
+
+    public User(long id, String username) {
+        this.id = id;
+        this.username = username;
+    }
+
+    public User(String username) {
+        this.username = username;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+}
